Add SpanEnumerable<T> and boxing
To avoid allocations, every NoAlloq operation returns a struct that includes in
its type a complete description of the operations performed so far. This makes
it hard to pass such a value as argument to another function, or to return it.

When this is necessary, NoAlloq now lets the user pay the cost for polymorphism
and type erasure by invoking the `.Box()` method. This returns an instance of
`SpanEnumerable<T>` where all details beyond the return type `T` have been
erased. This is implemented by:

 - boxing the internal structure that implements `IProducer<TIn, TOut>` or
   `IProducer<TOut>` to hide its concrete type, and
 - if there is an input span, using `MemoryMarshal.Cast` to convert that span
   to a `ReadOnlySpan<byte>` (and doing the reverse conversion when enumerating).
   This means, of course, that `.Box()` is only possible if the original span
   type was of an unmanaged type.