Make span-enumerables compatible with foreach
In C#, foreach is duck-typed: it expects the expression to have a public method
named `GetEnumerator()`, and the type returned by that method to have a public
property `Current` and a boolean-returning public method `MoveNext()`. This
works even if the enumerated or enumerator types are `ref struct` and so cannot
implement `IEnumerable<T>` or `IEnumerator<T>`.