Add OfType method and tests, fix null/default elements in CopyInto (#2)
Added OfType utility method in SpanEnumerable, OfTypeExtensions, and 
two producer classes, and provided corresponding tests. This method is 
meant to emulate the behavior of the .OfType<T>() LINQ method. It is 
used when we want to filter elements of a certain type from a span.

Also made a minor change in CopyInto method of OrderingExtensions to 
return the output after copying elements into the backing array. The 
change fix default/null elements in backing span.